--- .env
@@ -1,3 +1,0 @@
-CLIENT_ID="J-_gFicDvTwkRoBRR5TBwQ"
-YELP_API_KEY="6XAT9PezDIL-wN33pP9RGNJVMlg03CQAY2gJ4uFseU4Yr-2xa2OCybUe5zj1Q922UVAvJk8Niwoks8w0JYq5R7CXhxiDcEEMXNmbCq0cFVDScJXH4zYayPiGrNOCZ3Yx"
-YELP_API_URL = 'https://api.yelp.com/v3/businesses/search'
